+++ fabric/src/main/java/com/lx862/jcm/mod/scripting/jcm/JCMScripting.java
@@ -13,11 +13,13 @@ import com.lx862.mtrscripting.lib.org.mozilla.javascript.NativeJavaClass;
 import org.mtr.mod.client.MinecraftClientData;
 
 public class JCMScripting {
-    private static final ScriptManager scriptManager = new ScriptManager();
+    private static ScriptManager scriptManager;
     /**
      * Called once when the mod entrypoint is invoked
      */
     public static void register() {
+        if(scriptManager == null) scriptManager = new ScriptManager();
+        scriptManager.getClassShutter().setEnabled(!JCMClient.getConfig().disableScriptingRestriction);
         ScriptingAPI.registerAddonVersion("jcm", Constants.MOD_VERSION);
 
         scriptManager.getClassShutter().allowClass(ClassRule.parse("org.mtr.*"));