+++ fabric/src/main/java/com/lx862/mtrscripting/core/MTRClassShutter.java
@@ -1,5 +1,6 @@
 package com.lx862.mtrscripting.core;
 
+import com.lx862.jcm.mod.config.ClientConfig;
 import com.lx862.mtrscripting.api.ClassRule;
 import com.lx862.mtrscripting.lib.org.mozilla.javascript.ClassShutter;
 import it.unimi.dsi.fastutil.objects.ObjectArrayList;
@@ -11,6 +12,8 @@ import java.util.Arrays;
  * Limit the usage of arbitrary java class for security reason
  */
 public class MTRClassShutter implements ClassShutter {
+    /** Whether the class shutter is active or not. If false, it will allow all class through (Except itself) as if no restrictions is put in place. */
+    private boolean shutterEnabled;
     private final ObjectList<ClassRule> allowedScriptClasses = new ObjectArrayList<>();
     private final ObjectList<ClassRule> deniedScriptClasses = new ObjectArrayList<>();
 
@@ -36,6 +39,10 @@ public class MTRClassShutter implements ClassShutter {
         );
     }
 
+    public void setEnabled(boolean shutterEnabled) {
+        this.shutterEnabled = shutterEnabled;
+    }
+
     /**
      * By default, MTR Scripting does not allow loading any arbitrary java class for security reasons<br>
      * Here, you can explicitly allow a class to be loaded.<br>
@@ -52,15 +59,19 @@ public class MTRClassShutter implements ClassShutter {
 
     @Override
     public boolean visibleToScripts(String fullClassName) {
-        return fullClassName.startsWith("com.lx862.mtrscripting.util") || fullClassName.startsWith("com.lx862.mtrscripting.lib.org.mozilla") || isClassAllowed(fullClassName);
+        return !fullClassName.equals(MTRClassShutter.class.getName()) && !fullClassName.equals(ClientConfig.class.getName()) && isClassAllowed(fullClassName);
     }
 
-    private boolean isClassAllowed(String str) {
+    private boolean isClassAllowed(String className) {
+        if(!shutterEnabled) return true;
+        if(className.startsWith("com.lx862.mtrscripting.util")) return true;
+        if(className.startsWith("com.lx862.mtrscripting.lib.org.mozilla")) return true;
+
         for(ClassRule cs : deniedScriptClasses) {
-            if(cs.match(str)) return false;
+            if(cs.match(className)) return false;
         }
         for(ClassRule cs : allowedScriptClasses) {
-            if(cs.match(str)) return true;
+            if(cs.match(className)) return true;
         }
         return false;
     }